Rename passes for consistency.
