+++ test/fx/test_constructor_mover_pass.py
@@ -3,7 +3,7 @@
 import torch
 from torch.fx.experimental.proxy_tensor import make_fx
 from torch.fx.passes.infra.pass_base import PassResult
-from torch.fx.passes.move_constructors_pass import (
+from torch.fx.passes.constructor_mover_pass import (
     ConstructorMoverPass,
     ZeroOrMultipleDevicesError,
 )
